split graph_drawer. delete useless => change working with type in app.py
