+++ src/task2/graph_drawer.py
@@ -3,14 +3,8 @@ from PyQt5.QtWidgets import (QWidget, QLabel,
 from PyQt5.QtGui import QPainter, QBrush, QPen
 from PyQt5.QtCore import Qt, QPoint
 from data_models import Crossroad, Town, Market, Warehouse
+from drawer_utils import CustomLabel, Point
 
-
-buildings_to_info = {
-    0: ('Crossroad', Qt.gray),
-    1: ('Town', Qt.darkGreen),
-    2: ('Market', Qt.yellow),
-    3: ('Warehouse', Qt.darkRed),
-}
 
 buildings_to_type = {
     0: Crossroad,
@@ -84,6 +78,7 @@ class GraphDrawer(QWidget):
         self.update()
 
     # TODO, maybe later don't recreate new building, if type wasn't be changed
+    # TODO create objects outside
     def setBuildings(self, buildings):
         self.buildings = buildings
         for building in self.buildings['posts']:
@@ -163,82 +158,3 @@ class GraphDrawer(QWidget):
 
     def mouseReleaseEvent(self, QMouseEvent):
         self.pressing = False
-
-
-class CustomLabel(QLabel):
-    def __init__(self, text, parent, background_color=None):
-        super().__init__(text, parent)
-        if background_color:
-            self.setStyleSheet(
-                f"QLabel {{background-color: {background_color};}}")
-
-    def setFontSize(self, font_size):
-        font = self.font()
-        font.setPointSize(font_size)
-        self.setFont(font)
-        self.resize(self.sizeHint())
-
-
-class Point(QWidget):
-    def __init__(self, idx, radius, parent=None, font_size=None):
-        super().__init__(parent)
-
-        self.town_idx = idx
-        self._building = None
-        self.font_coefficient = 1.5
-        self.idx_label = CustomLabel(str(idx), self)
-        self._outer_radius = radius
-        self._inner_radius = int(radius * 0.8)  # ? property
-        self._font_size = font_size or int(1 / len(str(idx)) * self._inner_radius * self.font_coefficient)
-        self._scale = 0
-        self.thickness = 2
-        self.show()
-
-    @property
-    def outer_radius(self):
-        return int(self.scale * self._outer_radius) + 1
-
-    @property
-    def inner_radius(self):
-        return int(self.scale * self._inner_radius) + 1
-
-    @property
-    def font_size(self):
-        return int(self.scale * self._font_size) + 1
-
-    @property
-    def scale(self):
-        return self._scale
-
-    @scale.setter
-    def scale(self, value):
-        self._scale = value
-        self.resize((self.outer_radius + self.thickness + 1) * 2, (self.outer_radius + self.thickness + 1) * 2)
-
-    def paintEvent(self, e):
-        inner_radius = self.inner_radius
-        outer_radius = self.outer_radius
-        font_size = self.font_size
-
-        painter = QPainter()
-        painter.begin(self)
-
-        painter.setPen(QPen(Qt.black, int(self.scale * self.thickness)))
-
-        painter.drawEllipse(self.thickness, self.thickness, 2 * outer_radius, 2 * outer_radius)
-
-        painter.setBrush(self._building.color)
-
-        diff = outer_radius - inner_radius
-        painter.drawEllipse(diff + self.thickness, diff + self.thickness, 2 * inner_radius, 2 * inner_radius)
-
-        self.idx_label.setFontSize(font_size)
-        self.idx_label.move(outer_radius - self.idx_label.width() // 2, outer_radius - self.idx_label.height() // 2)
-
-        painter.end()
-
-    def setBuilding(self, building):
-        self._building = building
-        self.setToolTip(str(self._building))
-
-
